Customer Churn Banking/deployment/requirements.txt
streamlit == 1.38.0
pandas == 2.2.2
seaborn == 0.13.2
matplotlib == 3.8.4
numpy == 1.26.4
scikit-learn==1.5.1
Pillow 
plotly
imblearn